Link clang using ld.bfd instead of gold.
This is a workaround for
https://github.com/kraj/meta-clang/issues/119

There is a patch on that bug which modifies `gold` to write
interp sections of 0x1000 bytes, to avoid problems when relocating
binaries during SDK installs.  For reasons that aren't yet clear,
the binaries linked as part of `meta-clang` don't have this long
`interp` section.  This means that they can't be relocated if the
new (relocated) path to the loader is too long (which is if the
prefix to `sysroots` is longer than `/usr/local/oecore-x86_64`).

This seems to be either because

a) the patch to gold isn't working, or
b) the patched gold isn't being used to link clang